+++ wallet/src/main/java/at/fhv/enterpriseapplications/wallet/dto/WalletDto.java
@@ -1,0 +1,10 @@
+package at.fhv.enterpriseapplications.wallet.dto;
+
+import jakarta.validation.constraints.NotNull;
+
+import javax.swing.*;
+
+public record WalletDto(
+        @NotNull String username,
+        @NotNull Integer coins
+) {}